+++ src/services/AccessService.tsx
@@ -72,6 +72,7 @@ import { OSPRegister } from 'components/overlays/OSPRegister'
 import { OSPConsent } from 'components/overlays/OSPConsent'
 import { OSPRegisterNext } from 'components/overlays/OSPRegister/OSPRegisterNext'
 import CompanyCertificateDetails from 'components/overlays/CompanyCertificateDetails'
+import DeleteCompanyCertificateConfirmationOverlay from 'components/overlays/CompanyCertificateDetails/DeleteCompanyCertificateConfirmationOverlay'
 
 let pageMap: { [page: string]: IPage }
 let actionMap: { [action: string]: IAction }
@@ -216,6 +217,8 @@ export const getOverlay = (overlay: OverlayState) => {
       return <UpdateCertificate id={overlay.id} />
     case OVERLAYS.COMPANY_CERTIFICATE_DETAILS:
       return <CompanyCertificateDetails id={overlay.id} />
+    case OVERLAYS.COMPANY_CERTIFICATE_CONFIRM_DELETE:
+      return <DeleteCompanyCertificateConfirmationOverlay id={overlay.id} />
     default:
       return <NotFound />
   }