+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - UI components created to display filter, sort and company certificates
   - Company Certificate Details overlay
   - Upload new Company Certificate overlay
+  - Delete certificate screen
 - Company Roles
   - Fetch the standard library data from standards.json and implement Table component to display it in the company roles section.